+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v1.2
+
+- デプロイ枚数オプションの n, n_example をnullableに変更。nullのとき、全画像を出力。
+- オプション変更: `--makesample <N>` -> `--generate-samplefile`
+
 ## v1.1
 
 - 自作カラーマップをサポート。